Fix crash: abs(Int16.min) overflow in waveform extractor
abs(-32768) overflows Int16 (max is 32767), causing EXC_BREAKPOINT.
Widen to Int32 before abs() to handle the full Int16 range safely.

Found via crash log: Thread 5 crashed in
WaveformExtractor.extract → abs<A>(_:) → SignedNumeric.- prefix

+++ VideoEditor/Packages/EditorCore/Sources/EditorCore/Media/WaveformExtractor.swift
@@ -66,7 +66,8 @@ public struct WaveformExtractor: Sendable {
                 let timeInSeconds = Double(currentSampleIndex) / sampleRate
                 let bucketIndex = min(Int(timeInSeconds / bucketDuration), sampleCount - 1)
 
-                let amplitude = Float(abs(sample)) / Float(Int16.max)
+                // Use Int32 to avoid overflow on Int16.min (-32768)
+                let amplitude = Float(abs(Int32(sample))) / Float(Int16.max)
                 if amplitude > peaks[bucketIndex] {
                     peaks[bucketIndex] = amplitude
                 }
